@@ -5,7 +5,7 @@ author: Reno Rizky Ollantino <reno.rizky.ollantino@gmail.com>
 homepage: https://github.com/bleuxstrife/pit_sms_call_log
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
